+++ src/pyconfita/backend/vault/vault.py
@@ -22,7 +22,8 @@ class KeyRef:
     key: str
 
     def get_cache_key(self) -> str:
-        return f"{self.path}-{self.key}"
+        # return f"{self.path}-{self.key}"
+        return f"{self.key}"
 
 
 class Backend(_Backend):
@@ -61,7 +62,8 @@ class Backend(_Backend):
             raise Exception("Vault logger must not be None")
         self.logger = logger
         self.cache = None
-        if enable_cache:
+        self.enable_cache = enable_cache
+        if self.enable_cache:
             maxsize = kwargs.get("cache_maxsize", 1024)
             ttl = kwargs.get("cache_ttl", 600)  # Defaults to 10min
             self.cache = Cache(maxsize=maxsize, ttl=ttl)
@@ -252,7 +254,7 @@ class Backend(_Backend):
         """
         Cache key-value store (loaded from path) if caching is enabled.
         """
-        if self.cache is not None:
+        if self.enable_cache:
             for k, v in kv_store.items():
                 cache_key = KeyRef(path=path, key=k).get_cache_key()
                 try:
@@ -260,7 +262,7 @@ class Backend(_Backend):
                     self.logger.log(
                         **{
                             "level": "debug",
-                            "message": {"message": f"[Vault] Set key {k} in cache"},
+                            "message": {"message": f"[Vault] Set key {k}" f" in cache"},
                         }
                     )
                 except Exception as e:
@@ -269,7 +271,7 @@ class Backend(_Backend):
                             "level": "error",
                             "message": {
                                 "message": f"[Vault] Failed to cache value"
-                                f" for cache key={cache_key}"
+                                f" for key {k}"
                             },
                         }
                     )
@@ -291,7 +293,7 @@ class Backend(_Backend):
 
         _path = kwargs.get("path", self.default_key_path)
         k_ref = KeyRef(path=_path, key=key)
-        if self.cache is not None:  # Cache enabled
+        if self.enable_cache:
             self.logger.log(
                 **{
                     "level": "debug",
@@ -308,7 +310,9 @@ class Backend(_Backend):
                 self.logger.log(
                     **{
                         "level": "debug",
-                        "message": {"message": f"[Vault] key {key} not found in cache"},
+                        "message": {
+                            "message": f"[Vault] key {key} not" f" found in cache"
+                        },
                     }
                 )
                 kv_store = self._get_kv_store_when_ready(path=k_ref.path)
@@ -319,7 +323,8 @@ class Backend(_Backend):
                 **{
                     "level": "debug",
                     "message": {
-                        "message": f"[Vault][cache disabled] call _get_key_when_ready"
+                        "message": f"[Vault][cache disabled] call"
+                        f" _get_key_when_ready"
                     },
                 }
             )