+++ CHANGELOG.md
@@ -1,10 +1,14 @@
+## 1.0.11 (2022-10-25)
+
+- Fixed Vault backend caching
+
 ## 1.0.10 (2022-10-25)
 
 - Fixed Vault backend caching
 
 ## 1.0.9 (2022-10-25)
 
--
+- Fixed Vault backend caching
 
 ## 1.0.8 (2022-10-25)